// src/common/constants/response-messages.enum.ts
export enum RESPONSE_MESSAGES_CODE {
  LOGIN_FAIL = 'LOGIN_FAIL',
  CREATE_ERROR = 'CREATE_ERROR',
  EMAIL_EXIST = 'EMAIL_EXIST',
  INVALID = 'INVALID',
  ERROR = 'ERROR',
  EMAIL_SEND_FAIL = 'EMAIL_SEND_FAIL',
  NOT_FOUND = 'NOT_FOUND',
  SELF_DELETE = 'SELF_DELETE',
  SUPER_ADMIN_DELETE = 'SUPER_ADMIN_DELETE',
  COMPANY_EXIST = 'COMPANY_EXIST',
  ORDER_NOT_FOUND = 'ORDER_NOT_FOUND',
  DRIVER_NOT_FOUND = 'DRIVER_NOT_FOUND',
  TRUCK_NOT_FOUND = 'TRUCK_NOT_FOUND',
  PHONE_EXIST = 'PHONE_EXIST',
  EMAIL_EXISTED = 'EMAIL_EXISTED',
  TRUCK_OWNER_NOT_FOUND = 'TRUCK_OWNER_NOT_FOUND',
  ALREADY_FAVORITED = 'ALREADY_FAVORITED',
  CREATE_ORDER_ERROR = 'CREATE_ORDER_ERROR',
  NOT_ASSIGNED = 'NOT_ASSIGNED',
  DELETED_ACCOUNT = 'DELETED_ACCOUNT',
  DRIVER_ALREADY_ADDED = 'DRIVER_ALREADY_ADDED',
  KICKED_OUT = 'KICKED_OUT',
  WRONG_PHONE_NUMBER_FORMAT = 'WRONG_PHONE_NUMBER_FORMAT',
  PHONE_INPUT_TOO_LONG = 'PHONE_INPUT_TOO_LONG',
  ORDER_TAKEN = 'ORDER_TAKEN',
  TOKEN_ERROR = 'TOKEN_ERROR',
  ORDER_PERMISSION = 'ORDER_PERMISSION',
  LACK_DRIVER_OR_TRUCK = 'LACK_DRIVER_OR_TRUCK',
  FINDNEWTRUCK_ERROR_STATUS = 'FINDNEWTRUCK_ERROR_STATUS',
  NOT_VERIFIED = 'NOT_VERIFIED',
  FOLDER_NOT_EXISTED = 'FOLDER_NOT_EXISTED',
  COMPANY_NOT_EXISTED = 'COMPANY_NOT_EXISTED',
  DRIVER_EXISTED = 'DRIVER_EXISTED',
  CUSTOMER_CANCEL_ORDER = 'CUSTOMER_CANCEL_ORDER',
  DRIVER_CANCEL_ORDER = 'DRIVER_CANCEL_ORDER',
  ORDER_CANCELED = 'ORDER_CANCELED',
  EMAIL_NOT_VERIFY = 'EMAIL_NOT_VERIFY',
  PHONE_OTP_VERIFY = 'PHONE_OTP_VERIFY',
  EMAIL_OR_PHONE_NOT_VERIFY = 'EMAIL_OR_PHONE_NOT_VERIFY',
  INCORRECT_OTP = 'INCORRECT_OTP',
  OTP_EXPIRED = 'OTP_EXPIRED',
  TOKEN_EXPIRED = 'TOKEN_EXPIRED',
  PHONE_NOT_EXIST = 'PHONE_NOT_EXIST',
  PHONE_ALREADY_VERIFIED = 'PHONE_ALREADY_VERIFIED',
  PICKUP_CODE_INCORRECT = 'PICKUP_CODE_INCORRECT',
  DELIVERY_CODE_INCORRECT = 'DELIVERY_CODE_INCORRECT',
  ADMIN_NOT_FOUND = 'ADMIN_NOT_FOUND',
  CUSTOMER_NOT_FOUND = 'CUSTOMER_NOT_FOUND',
  EMPLOYEE_NOT_FOUND = 'EMPLOYEE_NOT_FOUND',
  EMAIL_NOT_FOUND = 'EMAIL_NOT_FOUND',
  SERVICE_TYPE_OR_ORDER_TYPE = 'SERVICE_TYPE_OR_ORDER_TYPE',
  SERVICE_TYPE = 'SERVICE_TYPE',
  DETAIL_REQUIRED = 'DETAIL_REQUIRED',
  CARGO_TYPE = 'CARGO_TYPE',
  CARGO_TYPE_OR_NAME = 'CARGO_TYPE_OR_NAME',
  TRUCK_SPECIAL_OR_QUANTITY = 'TRUCK_SPECIAL_OR_QUANTITY',
  CONTAINER_PRICE_REQUIRED = 'CONTAINER_PRICE_REQUIRED',
  CONTAINER_REQUIRED = 'CONTAINER_REQUIRED',
  SUPPER_EXIST = 'SUPPER_EXIST',
  CITY_NOT_FOUND = 'CITY_NOT_FOUND',
  UNAUTHORIZED = 'UNAUTHORIZED',
  BANK_ACCOUNT_EXIST = 'BANK_ACCOUNT_EXIST',
  BANK_ACCOUNT_NOT_EXISTED = 'BANK_ACCOUNT_NOT_EXISTED',
  LIMIT_ORDER = 'LIMIT_ORDER',
  FAVORITE_TRUCK_INCORRECT = 'FAVORITE_TRUCK_INCORRECT',
  FILE_FORMAT = 'FILE_FORMAT',
  PICKUP_ADDRESS_NOT_FOUND = 'PICKUP_ADDRESS_NOT_FOUND',
  PICKUP_DROPOFF_NOT_FOUND = 'PICKUP_DROPOFF_NOT_FOUND',
  CANNOT_UPDATE_SYNC_DATA = 'CANNOT_UPDATE_SYNC_DATA',
  SYNC_DATA_IS_PROCESSING = 'SYNC_DATA_IS_PROCESSING',
  INCORRECT_CURRENT_PASSWORD = 'INCORRECT_CURRENT_PASSWORD',
}