// src/common/constants/black-box.enum.ts
export enum BLACK_BOX_TYPE {
  DINH_VI_HOP_QUY = 'DINH_VI_HOP_QUY',
  ANTECK = 'ANTECK',
  VIETMAP = 'VIETMAP',
}

export enum VEHICLE_TYPE {
  TRUCK = 'truck',
  SHIP = 'ship',
}

export enum VEHICLE_STATUS {
  UNKNOWN = -1,
  RUNNING = 0,
  STOP = 1,
  LOST_SIGNAL = 2,
}

export enum DISPLAY_ON {
  CUSTOMER = 'CUSTOMER',
  TRUCKOWNER = 'TRUCKOWNER',
}